backend/requirements.txt
# --- Django Core ---
Django==4.2.16
django-environ==0.11.2

# --- Base de Datos ---
psycopg==3.2.3

# --- API & REST ---
djangorestframework==3.15.2
djangorestframework-simplejwt
djangorestframework-simplejwt[crypto]
django-cors-headers
django-filter
python-dateutil

# --- Utilidades y Herramientas de Desarrollo ---
pre-commit
black
isort
flake8
django-extensions
django-debug-toolbar
django-silk
django-soft-delete

# --- Tareas Asíncronas y Cache ---
redis
celery
pyotp

# --- Imágenes ---
Pillow
cryptography
pyjwt
    django-phonenumber-field[phonenumbers]
